requirements.txt
streamlit>=1.28.0
folium>=0.14.0
streamlit-folium>=0.15.0
pandas>=2.0.0
numpy>=1.24.0
requests>=2.28.0
python-dotenv>=1.0.0
loguru>=0.7.0
